fixed bad-conjecture-type bug
git-svn-id: https://svn-i2.informatik.rwth-aachen.de/repos/libalf/trunk@949 ee1db7db-6c55-0410-a809-e4d5348c71a5

+++ libalf/dispatcher/src/co_learning_algorithm.cpp
@@ -250,7 +250,7 @@ bool co_learning_algorithm::handle_command(int command, basic_string<int32_t> & 
 			} else {
 				if(!this->sv->client->stream_send_int(1))
 					return false;
-				if(!this->sv->client->stream_send_int(cj->get_type()))
+				if(!this->sv->client->stream_send_int( (int) cj->get_type() ))
 					return false;
 
 				serial = cj->serialize();
